backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5
django-extensions==3.2.3
